project/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
Django==5.1.2
django-allauth==65.0.2
djangorestframework==3.15.2
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
idna==3.10
promise==2.3
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
urllib3==2.2.3
